This is a work-in-progress commit before changing all of the core types to interfaces (and enums) as detailed here in the notes: http://www.wrml.org/docs/notes/wrml-java-api.txt
I've been working on laying the groundwork for WRML Java API.
Specifically, I have been building the interplay between fields, links,
all of the metadata types, the auto-generation foundations, etc. 

Now, I've reached the point where I want to start building WRML with
WRML. I was working on the Service Java API and thinking that it might
nicely serve as a common abstraction for both the client's remote
REST-based communications and the server's relationship to its backend
connection. As the design stands now, the Service Java interface is
intended to serve both needs, with the Context interface serving as a
bridge which can vary as needed to support the differing runtime needs
of a client and a server. The missing piece is that all of WRML's core
types need to become WRML Models. My proposed design approach for this
bootstrap can be found in the notes linked above.
 
This is a worthwhile goal because I know I want us to easily write
Eclipse and IntelliJ plug-ins that share a common Java client core,
which happens to be the same core that is used by the web resource
server engine. This recursive or fractal nature of WRML is all about
CRUD. Regardless of what part of the application's stack needs data, the
interactions are web-oriented using Models, Fields, and Links.

BTW Brian, regarding the IDE plug-ins I thought it would be fun to each
pick one (to mirror our IDE choices), so I would build the WRML eclipse
tools and you would do the IntelliJ ones. And we will have a friendly
coopetition to see who can make the coolest WRML tools. =)

Again, my next step is to strip/modify all of the core types so that
they look like the interfaces and enums detailed here in the notes: 

http://www.wrml.org/docs/notes/wrml-java-api.txt