Commit after changing all of the core types to interfaces (and enums) as detailed here in the notes: http://www.wrml.org/docs/notes/wrml-java-api.txt
Along the way, I repackaged most of the code again into a model base
package, which is meant to callout that these classes are all meant to
be replaced by WRML's eventual automation.

My next step is to see how much of the core code can be cleanly written
and organized within the root runtime classes (Model, Link, Context, and
Services).

I will probably start to explore the baseclass Proxy solution so that
the core WRML model interfaces can be start to be instantiated as
Models.