The service framework is starting to take shape. The basic premise is a layered system with each service implementing an augmented Map interface which mirrors the capabilities available on most storage abstractions, including web services and HTTP's methods in general. The service delegation (proxy and multiproxy) can be used to model a linked list of services that may even be used to mask remote communications.
Implemented the Proxy approach to addressing RuntimeModels as Model
interfaces. The App's test program seems to be sort of working although
all of the code is still very proto/poc and fragile as a
result. Specifically, the framework code seems to have a few stack over
flow bugs lurking somewhere to be debugged.

The Context and Service interfaces and the connection between them is
not sitting very well with me right now. I think that there is a
cleaner design waiting to emerge.

The WebService is just a placeholder right now but my thought is that it
will be the client side's main interface to REST-remote data.