Continued to refactor the base model and type system.
Added a Native type to allow for native Objects to be stored in Model
fields.

Reworked the design of the static interface so it is now lazily
available from the inherently dynamic Model.

Added some key Constraint types to allow for parameterized Schemas. 

Still in the process of designing link clicking.