Repackaged the code to make it clear which parts are for the runtime system versus part of WRML's exposed Java API.
Started to work on the runtime model for REST APIs. Specifically the
connection that each model has with its "origin" ResourceTemplate and
how that connection allows hypermedia links to be represented via Java
methods. So that a "team" link relation becomes a function called
"getTeam". This is the lazy loading part of WRML that uses the WWW as
its database. This is the "MWM" (Model-Web Mapping), which is like the
WRML equivalent of ORM (Object-Relational Mapping).