Reached a point of stability that seems to prove that the core bootstrapping model will work as designed.
In its current state, the core's runtime system assumes that a schema
"compiler" has already run, meaning that the core itself makes heavy use
of statically referenced WRML schemas (org.wrml.Model
sub-interfaces). A next phase will be to build the compiler to translate
the WRML schema definition "syntax" into a Java interface (a .class
file). For example this code will need to translate type constraints
into parameterized types and Text field syntax constraints like
"URI" into text-transformed field "getter" method return types
like java.net.URI.

The code is obviously in need of Log4J. So there is another task.

The ModelGraph structure is not yet all it is meant to be. It will
eventually contain the in-memory representation of a de-serialized
Model's hierarchical document model (sort of like a browser's DOM). The
structure represents the nesting scoping of embedded models. For
examples the JSON document: 

{
  "childA" : {
  },
  "childB" : {
  }
}  

Would result in a Model graph that is rooted by a Model representing the
first/top '{' which has two child Models (from field relationships). 

This graph will be interrogated immediately following (or during) its
initialization to ensure Model cache consistency whenever a Document is
redundantly read within the same Document (graph loops) or is simply a
double of some Model that we have already cached. In either case this
ModelGraph data structure can be used to help ensure the single-ness or
one-to-one-ness of our cached Models to their associated server-owned
resources.

Once the graph is completed, another task would be to expose it via JMX
Mbeans.